style: rename styles files
